requirements.txt
Flask==3.1.0
tensorflow==2.18.0
Pillow==11.0.0
gunicorn==23.0.0
numpy==2.0.2
tensorflow-intel==2.18.0
tensorboard==2.18.0
